+++ gradle-plugin/src/main/kotlin/proguard/gradle/plugin/android/tasks/CollectConsumerRulesTask.kt
@@ -2,9 +2,12 @@ package proguard.gradle.plugin.android.tasks
 
 import java.io.File
 import java.io.FileOutputStream
+import java.io.Serializable
 import org.gradle.api.DefaultTask
+import org.gradle.api.GradleException
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.ResolvedArtifact
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
@@ -16,15 +19,27 @@ open class CollectConsumerRulesTask : DefaultTask() {
     @get:OutputFile
     lateinit var outputFile: File
 
+    @get:Input
+    var consumerRuleFilter: List<ConsumerRuleFilterEntry> = emptyList()
+
     @TaskAction
     fun extractConsumerRules() {
         if (outputFile.exists()) {
             outputFile.delete()
         }
+        logger.debug("Writing out consumer rules to '$outputFile'...")
         outputFile.createNewFile()
 
+        val matchedConsumerRuleFilterEntries = mutableListOf<ConsumerRuleFilterEntry>()
         val resolvedArtifacts = consumerRulesConfiguration.resolvedConfiguration.resolvedArtifacts
-        resolvedArtifacts.forEach { artifact ->
+
+        for (artifact in resolvedArtifacts) {
+            val entry = ConsumerRuleFilterEntry(artifact.moduleVersion.id.group, artifact.moduleVersion.id.name)
+            if (entry in consumerRuleFilter) {
+                logger.debug("Skipping consumer rules of '${artifact.file}'...")
+                matchedConsumerRuleFilterEntries += entry
+                continue
+            }
             project.fileTree(artifact.file).forEach { file ->
                 val relativeFilePath = file.absolutePath.removePrefix(artifact.file.absolutePath)
                 file.inputStream().bufferedReader().use { reader ->
@@ -65,5 +80,12 @@ open class CollectConsumerRulesTask : DefaultTask() {
                 }
             }
         }
+
+        if (matchedConsumerRuleFilterEntries.size < consumerRuleFilter.size) {
+            val notMatchedConsumerRuleFIlterEntries = consumerRuleFilter - matchedConsumerRuleFilterEntries
+            throw GradleException("Consumer rule filter entr${if (notMatchedConsumerRuleFIlterEntries.size > 1) "ies" else "y"} '${notMatchedConsumerRuleFIlterEntries.joinToString { "${it.group}:${it.module}" }}' did not match any dependency")
+        }
     }
 }
+
+data class ConsumerRuleFilterEntry(val group: String, val module: String) : Serializable